pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "authlib"
version = "1.0.0"
description = "A scalable, framework-agnostic Python authentication library with JWT token management and password reset functionality"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AuthLib Contributors", email = "info@authlib.dev"}
]
keywords = [
    "authentication",
    "authorization",
    "jwt",
    "signup",
    "login",
    "password-reset",
    "oauth",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
]

dependencies = [
    "sqlalchemy>=2.0.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

fastapi = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

flask = [
    "flask>=3.0.0",
]

psycopg = [
    "psycopg[binary]>=3.1.0",
]

[project.urls]
"Homepage" = "https://github.com/authlib/authlib-python"
"Documentation" = "https://github.com/authlib/authlib-python#readme"
"Repository" = "https://github.com/authlib/authlib-python.git"
"Issues" = "https://github.com/authlib/authlib-python/issues"
"Changelog" = "https://github.com/authlib/authlib-python/blob/main/CHANGELOG.md"
"PyPI" = "https://pypi.org/project/authlib/"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
